chore: add new module for keplrStrategy and improve types
